Added dependencies for Arduino Lib-Manager
version 1.0.1:
Added dependencies for Arduino Lib-Manager like keywords, library.properties etc.
