+++ library.properties
@@ -1,0 +1,9 @@
+name=Arduino-MAX17055_Driver
+version=1.0.0
+author=Awot Ghirmai <ghirmai.awot@gmail.com>
+maintainer=Awot Ghirmai <ghirmai.awot@gmail.com>
+sentence=An Arduino library for the Maxim Fuel Gauge MAX17055
+paragraph=The library provides functions to read and write registers of the fuel gauge and initializes the fuel gauge with correct values for further reading by the host uC. See example code and documentation for further details.
+category=Other
+url=https://github.com/AwotG/Arduino-MAX17055_Driver
+architectures=*
